requirements.txt
pandas==2.0.3
openpyxl==3.1.2
pyyaml==6.0.1
eth-account==0.9.0
mnemonic==0.20
base58==2.1.1
cryptography==41.0.1
numpy==1.24.3

